move some graphQL files and change structure of models
