+++ Projects/Contracts/GraphQL/Giraffe/Contracts.graphql
