+++ Projects/Offer/Sources/OfferStore.swift
@@ -5,6 +5,7 @@ import Presentation
 import StoreKit
 import hAnalytics
 import hCore
+import hCoreUI
 import hGraphQL
 
 public struct OfferState: StateProtocol {
