+++ Projects/Payment/Project.swift
@@ -4,7 +4,7 @@ import ProjectDescriptionHelpers
 let project = Project.framework(
     name: "Payment",
     targets: Set([.framework, .tests, .example, .testing]),
-    projects: ["hCore", "hCoreUI"],
+    projects: ["hCore", "hCoreUI", "Contracts"],
     dependencies: ["CoreDependencies", "ResourceBundledDependencies"],
     sdks: [],
     includesGraphQL: true
