+++ Projects/hCore/Sources/PillowType.swift
@@ -1,0 +1,14 @@
+public enum PillowType: Codable {
+    case accident
+    case car
+    case cat
+    case dog
+    case home
+    case homeOwner
+    case pet
+    case rental
+    case student
+    case travel
+    case villa
+    case unknown
+}
