+++ Projects/hCoreUI/Sources/InsurableLimits/InsurableLimits.swift
@@ -48,3 +48,25 @@ public struct InsurableLimitsSectionView: View {
         .sectionContainerStyle(.transparent)
     }
 }
+
+public struct InsurableLimits: Codable, Hashable {
+    public let label: String
+    public let limit: String
+    public let description: String
+
+    public init(
+        fragment: GiraffeGraphQL.InsurableLimitFragment
+    ) {
+        label = fragment.label
+        limit = fragment.limit
+        description = fragment.description
+    }
+
+    init(
+        _ data: OctopusGraphQL.ProductVariantFragment.InsurableLimit
+    ) {
+        label = data.label
+        limit = data.limit
+        description = data.description
+    }
+}
